backend/requirements.txt
fastapi
pydantic>=2.0.0
requests==2.31.0
python-dotenv==1.0.1
python-whois==0.8.0
beautifulsoup4==4.12.3
tld==0.13.0
uvicorn==0.29.0
pytest==8.0.0
pytest-asyncio==0.23.5
pytest-cov==4.1.0
pytest-mock==3.12.0
httpx==0.26.0

